requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf_yasg==1.21.9
python-decouple==3.8
